createUsersAndLogin

+++ app/Http/Controllers/PostController.php
@@ -3,11 +3,16 @@
 namespace App\Http\Controllers;
 
 use App\Models\Post;
+use App\Models\User;
 use Illuminate\Http\Request;
-use tidy;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use Illuminate\Support\Facades\Auth;
+
+
 
 class PostController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
@@ -136,6 +141,7 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        $this->authorize("manegeUser" , User::class);
         $post->delete();
         return redirect()->route('posts.index');
     }
